requirements.txt
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
bcrypt==4.1.2
certifi==2022.6.15
cffi==1.16.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
coverage==7.4.4
cryptography==41.0.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.5
Django==5.0.2
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-import-export==3.3.7
django-redis==5.4.0
django-rest-passwordreset==1.4.0
django-rq==2.10.1
django-simple-history==3.4.0
django-storages==1.14.3
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.4.2
iniconfig==2.0.0
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
MarkupPy==1.14
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.3
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
paramiko==3.4.0
parso==0.8.3
pillow==10.2.0
platformdirs==2.5.2
pluggy==1.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyNaCl==1.5.0
pytest==8.1.1
pytest-cov==5.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.2
rq==1.15.1
rq-win @ git+https://github.com/michaelbrooks/rq-win.git@c6e65e38f8a6ca99a0c8e0dd6fe35f9b92837cca
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.3
tablib==3.5.0
times==0.7
tornado==6.4
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
tzdata==2022.1
virtualenv==20.15.1
virtualenv-clone==0.5.7
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0